move assert_positions to main package so can be used in debugging
