+++ mofun/helpers.py
@@ -92,3 +92,43 @@ def typekey(tup):
     if tuple(rev) <= tuple(tup):
         return tuple(rev)
     return tuple(tup)
+
+def assert_structure_positions_are_unchanged(orig_structure, final_structure, max_delta=1e-5, verbose=False):
+    return assert_positions_are_unchanged(orig_structure.positions, final_structure.positions, max_delta, verbose)
+
+def assert_positions_are_unchanged(p, new_p, max_delta=1e-5, verbose=False):
+    p_ordered = p[np.lexsort((p[:,0], p[:,1], p[:,2]))]
+    new_p_ordered = new_p[np.lexsort((new_p[:,0], new_p[:,1], new_p[:,2]))]
+    p_unmatched = []
+    p_corresponding = []
+    distances = np.full(len(p), max(9.99, 9.99 * max_delta))
+    for i, p1 in enumerate(p_ordered):
+        found_match = False
+        for j, p2 in enumerate(new_p_ordered):
+            # print(p2, p1, norm(np.array(p2) - p1))
+            if p2[2] - p1[2] > 1:
+                p_unmatched.append(p1)
+                break
+            elif (np21 := norm(np.array(p2) - p1)) < max_delta:
+                found_match = True
+                p_corresponding.append(new_p_ordered[j, :])
+                new_p_ordered = np.delete(new_p_ordered, j, axis=0)
+                distances[i] = np21
+                break
+        if not found_match:
+            p_corresponding.append([])
+
+    distances = np.array(distances)
+    if verbose:
+        for i, p1 in enumerate(p_ordered):
+            annotation = ""
+            if distances[i] > max_delta:
+                annotation = " * "
+            print(i, p1, p_corresponding[i], distances[i], annotation)
+        print("UNMATCHED coords in old positions: ")
+        for p1 in p_unmatched:
+            print(p1)
+        print("UNMATCHED coords in new positions: ")
+        for p1 in new_p_ordered:
+            print(p1)
+    assert (distances < max_delta).all()
